refactor(init): move ttObject to a separate file
